Started refactoring processors to separate files, added abstract base class
